# api/openapi/order.yml
openapi: 3.0.0
info:
  title: order service
  description: order service
  version: 1.0.0
servers:
  - url: 'https://{hostname}/api'
    variables:
      hostname:
        default: 127.0.0.1
paths:
  /customer/{customerID}/orders/{orderID}:
    get:
      description: "get order"
      parameters:
        - in: path
          name: customerID
          schema:
            type: string
          required: true

        - in: path
          name: orderID
          schema:
            type: string
          required: true

      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orders:
    post:
      description: "create order"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Order:
      type: object
      required:
        - id
        - customerID
        - status
        - items
        - paymentLink
      properties:
        id:
          type: string
        customerID:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        paymentLink:
          type: string

    Item:
      type: object
      required:
        - id
        - name
        - quantity
        - priceID
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32
        priceID:
          type: string

    CreateOrderRequest:
      type: object
      required:
        - customerID
        - items
      properties:
        customerID:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemWithQuantity"

    ItemWithQuantity:
      type: object
      required:
        - id
        - quantity
      properties:
        id:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1

    Error:
      type: object
      properties:
        message:
          type: string